Added .gitignore, refactored monopod env code (MonopodBase base-class for easier testing changes)
